Allow simulators other than GHDL and ModelSim in run.py
If the VUNIT_SIMULATOR environment variable is already set then use that
instead of overriding it.  This allows the regression tests to run with
the VHDL simulator I am developing at nickg/nvc.

As far as I know it's fine to set simulator-specific VUnit
options (e.g. ghdl.elab_flags) when that simulator is not selected.